Changes to 3 source members to stop a warning (Windows build) and 2 errors (wxWidgets build) with VS2013 Update project files to set WINVER=0x0501 for Windows XP (oldest currently supported PWS OS). Required for VS2013. Note current value (0x0500) means Windows 2000. Change Windows wxWidgets pwsafe project name to pwsafe11 in keeping with all other Windows wxWidgets project naming. Add YubiKey library to Windows wxWidgets build.

+++ src/ui/Windows/CompareResultsDlg.cpp
@@ -1446,7 +1446,7 @@ bool CCPListCtrl::IsSelected(DWORD_PTR iRow)
   POSITION pos = GetFirstSelectedItemPosition();
 
   while (pos) {
-    if (GetNextSelectedItem(pos) == iRow)
+    if (GetNextSelectedItem(pos) == (int)iRow)
       return true;
   }
